+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+- Add `:path` option to `telegram_webhook` route helper.
+- __Breaking change!__ Default route is generated using hashed bot token.
+  Please reconfigure webhook after update (`rake telegram:bot:set_webhook`).
+
 # 0.14.4
 
 - Update to Bot API 4.7